Moved compile_and_execute to easy_api example
